avoid reserving from the connection pool
This is not necessary for string escaping and will cause multiple Fibers
sharing a same visitor not designed to handle concurrency as string
encoding is delegated to the connection, which is a ConnectionPool that
will acquire a real_connection to be able to delegate the work.

Should there be less real connections than Fibers, this will cause
yielding while generating the sql using the visitor.

The visitor does not support being used by concurrent threads,
generating cast errors as values are prepared for the wrong column type